Orientation bugfix
Fixed a nasty bug where ScalingViewport was incorrectly calculating
viewportToPixel() and was swapping the positive and negative
directions.

For whatever reason, I had not noticed this bug before and instead
littered workarounds throughout the rest of the project.

All the traces of this bug *should* be stamped out now, and all my
test cases are performing in a sane manner. Plus, now the default
ScalingViewport renders in an identical manner to a vanilla
Viewport. This should be solved completely.